[feat/board-v4] multer를 통해서 이미지 올릴 수 있게 함

+++ controllers/comment.controller.js
@@ -41,6 +41,8 @@ class CommentController {
   async deleteComment(req, res, next) {
     try {
       const id = req.params.id;
+      if ((await commentService.getCommentUidByCid(id)) !== req.user.id)
+        throw new Error("permission denied");
       const comment = await commentService.deleteComment(id);
       res.status(200).json(comment);
     } catch (err) {
